Re-Format europarl recipe
