+++ egs3/europarl/s2t1/conf/inference_beam1.yaml
@@ -1,0 +1,47 @@
+recipedir: .
+expdir: exp/train_ctc
+decode_dir: ${expdir}/decode_debug
+
+model:
+  _target_: espnet2.bin.asr_inference.Speech2Text
+  asr_train_config: ${expdir}/config.yaml
+  asr_model_file: ${expdir}/last.ckpt
+  beam_size: 1
+
+dataset:
+  _target_: espnet3.data.DataOrganizer
+  test:
+    - name: test-clean
+      dataset:
+        _target_: egs3.librispeech_100.asr1.dataset.LibriSpeechDataset
+        split: test-clean
+    - name: test-other
+      dataset:
+        _target_: egs3.librispeech_100.asr1.dataset.LibriSpeechDataset
+        split: test-other
+
+  preprocessor:
+    _target_: espnet2.train.preprocessor.CommonPreprocessor
+    train: false
+    token_type: bpe
+    token_list: ${recipedir}/sentencepiece_model/tokens.txt
+    bpemodel: ${recipedir}/sentencepiece_model/bpe.model
+    text_cleaner:
+
+parallel:
+  env: slurm
+  n_workers: 4
+  options:
+    account: bbjs-delta-gpu
+    queue: gpuA40x4
+    cores: 1
+    processes: 1
+    memory: 16GB
+    death_timeout: 180
+    # walltime: 30:00
+    job_extra_directives:
+      - "--gres=gpu:1"
+      - "--time 5:00:00"
+      - "--ntasks-per-node=1"
+      - "--cpus-per-task=8"
+      - "--output=parallel_log/%j-%x.log"
