+++ egs3/europarl/s2t1/conf/owsm_finetune_gpu2.yaml
@@ -1,0 +1,8 @@
+# load default config
+defaults:
+  - train.yaml
+
+num_device: 2
+
+trainer:
+  strategy: ddp_find_unused_parameters_true
