--- egs3/europarl/s2t1/config.yaml
@@ -1,311 +1,0 @@
-recipedir: .
-num_device: ${oc.env:N_GPU}
-expdir: ${recipedir}/exp/gpu${num_device}
-statsdir:  ${recipedir}/exp/stats
-vocab_size: 50002
-seed: 2024
-
-parallel:
-  env: slurm
-  n_workers: 16
-  options:
-    account: bbjs-delta-cpu
-    queue: cpu
-    cores: 1
-    processes: 1
-    memory: 16GB
-    # walltime: 3:00:00
-    death_timeout: 180
-    job_extra_directives:
-      # - "--gres=gpu:1"
-      - "--time 1:00:00"
-      - "--ntasks-per-node=1"
-      - "--cpus-per-task=8"
-      - "--output=parallel_log/%j-%x.log"
-
-dataset:
-  _target_: espnet3.data.DataOrganizer
-  train:
-    # ASR dataset
-    - name: de_asr_train
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlASRDataset
-        split: de_en_train
-
-    - name: fr_asr_trian
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlASRDataset
-        split: fr_en_train
-
-    - name: it_asr_train
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlASRDataset
-        split: it_en_train
-
-
-    # ST dataset
-    - name: de_fr_train
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: de_fr_train
-
-    - name: de_it_train
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: de_it_train
-
-    - name: fr_de_train
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: fr_de_train
-
-    - name: fr_it_train
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: fr_it_train
-
-    - name: it_de_train
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: it_de_train
-
-    - name: it_fr_train
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: it_fr_train
-
-
-  valid:
-    # ASR data
-    - name: de_asr_dev
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlASRDataset
-        split: de_en_dev
-
-    - name: fr_asr_dev
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlASRDataset
-        split: fr_en_dev
-
-    - name: it_asr_dev
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlASRDataset
-        split: it_en_dev
-
-
-    # ST data
-    - name: de_fr_dev
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: de_fr_dev
-
-    - name: de_it_dev
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: de_it_dev
-
-    - name: fr_de_dev
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: fr_de_dev
-
-    - name: fr_it_dev
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: fr_it_dev
-
-    - name: it_de_dev
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: it_de_dev
-
-    - name: it_fr_dev
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: it_fr_dev
-
-
-  test:
-    # ASR data
-    - name: de_asr_test
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlASRDataset
-        split: de_en_test
-
-    - name: fr_asr_test
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlASRDataset
-        split: fr_en_test
-
-    - name: it_asr_test
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlASRDataset
-        split: it_en_test
-
-
-    # ST data
-    - name: de_fr_test
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: de_fr_test
-
-    - name: de_it_test
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: de_it_test
-
-    - name: fr_de_test
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: fr_de_test
-
-    - name: fr_it_test
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: fr_it_test
-
-    - name: it_de_test
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: it_de_test
-
-    - name: it_fr_test
-      dataset:
-        _target_: egs3.europarl.asr_st.dataset.EuroparlSTDataset
-        split: it_fr_test
-
-  preprocessor:
-    _target_: egs3.europarl.asr_st.dataset.OWSMTokenizeTransform
-    model_tag: espnet/owsm_v3.1_ebf
-
-model:
-  _target_: egs3.europarl.asr_st.model.OWSMFinetune
-  model_tag: espnet/owsm_v3.1_ebf
-
-optim:
-  _target_: torch.optim.AdamW
-  lr: 0.00005
-  weight_decay: 0.000001
-
-scheduler:
-  _target_: espnet2.schedulers.warmup_lr.WarmupLR
-  warmup_steps: 6000
-
-dataloader:
-  collate_fn:
-    _target_: espnet2.train.collate_fn.CommonCollateFn
-    int_pad_value: -1
-  train:
-    iter_factory:
-      _target_: espnet2.iterators.sequence_iter_factory.SequenceIterFactory
-      shuffle: true
-      collate_fn: ${dataloader.collate_fn}
-      num_workers: 4
-      num_iters_per_epoch: 15000
-      batches:
-        _target_: espnet2.samplers.build_batch_sampler.build_batch_sampler
-        type: numel
-        shape_files:
-          - ${statsdir}/train/feats_shape
-        batch_size: 4
-        batch_bins: 500000
-  valid:
-    iter_factory:
-      _target_: espnet2.iterators.sequence_iter_factory.SequenceIterFactory
-      shuffle: false
-      collate_fn: ${dataloader.collate_fn}
-      num_workers: 4
-      batches:
-        _target_: espnet2.samplers.build_batch_sampler.build_batch_sampler
-        type: numel
-        shape_files:
-          - ${statsdir}/valid/feats_shape
-        batch_size: 4
-        batch_bins: ${dataloader.train.iter_factory.batches.batch_bins}
-
-
-best_model_criterion:
-  - [valid/loss, 3, min]
-
-trainer:
-  accelerator: gpu
-  devices: ${num_device}
-  num_nodes: 1
-  accumulate_grad_batches: 1
-  gradient_clip_val: 1.0
-  log_every_n_steps: 500
-  max_epochs: 30
-  precision: bf16
-
-  logger:
-    - _target_: lightning.pytorch.loggers.TensorBoardLogger
-      save_dir: ${expdir}/tensorboard
-      name: tb_logger
-
-    - _target_: lightning.pytorch.loggers.WandbLogger
-      project: ESPnet3 development
-      save_dir: ${expdir}/wandb
-      name: Europarl
-
-  strategy: auto
-
-# fit:
-#   ckpt_path: ${recipedir}/exp/gpu1/last.ckpt
-
-test:
-  - name: de_asr_test
-    metrics:
-      - _target_: espnet3.metrics.WER
-        inputs:
-          ref: text
-          hyp: hypothesis
-  - name: fr_asr_test
-    metrics:
-      - _target_: espnet3.metrics.WER
-        inputs:
-          ref: text
-          hyp: hypothesis
-  - name: it_asr_test
-    metrics:
-      - _target_: espnet3.metrics.WER
-        inputs:
-          ref: text
-          hyp: hypothesis
-
-  # ST data
-  - name: de_fr_test
-    metrics:
-      - _target_: espnet3.metrics.BLEU
-        inputs:
-          ref: text
-          hyp: hypothesis
-  - name: de_it_test
-    metrics:
-      - _target_: espnet3.metrics.BLEU
-        inputs:
-          ref: text
-          hyp: hypothesis
-  - name: fr_de_test
-    metrics:
-      - _target_: espnet3.metrics.BLEU
-        inputs:
-          ref: text
-          hyp: hypothesis
-  - name: fr_it_test
-    metrics:
-      - _target_: espnet3.metrics.BLEU
-        inputs:
-          ref: text
-          hyp: hypothesis
-  - name: it_de_test
-    metrics:
-      - _target_: espnet3.metrics.BLEU
-        inputs:
-          ref: text
-          hyp: hypothesis
-  - name: it_fr_test
-    metrics:
-      - _target_: espnet3.metrics.BLEU
-        inputs:
-          ref: text
-          hyp: hypothesis
