+++ egs3/europarl/s2t1/dataset.yaml
@@ -1,0 +1,155 @@
+
+dataset:
+  _target_: espnet3.data.DataOrganizer
+  train:
+    # ASR dataset
+    - name: de_asr_train
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlASRDataset
+        split: de_en_train
+
+    - name: fr_asr_trian
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlASRDataset
+        split: fr_en_train
+
+    - name: it_asr_train
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlASRDataset
+        split: it_en_train
+
+
+    # ST dataset
+    - name: de_fr_train
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: de_fr_train
+
+    - name: de_it_train
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: de_it_train
+
+    - name: fr_de_train
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: fr_de_train
+
+    - name: fr_it_train
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: fr_it_train
+
+    - name: it_de_train
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: it_de_train
+
+    - name: it_fr_train
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: it_fr_train
+
+
+  valid:
+    # ASR data
+    - name: de_asr_dev
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlASRDataset
+        split: de_en_dev
+
+    - name: fr_asr_dev
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlASRDataset
+        split: fr_en_dev
+
+    - name: it_asr_dev
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlASRDataset
+        split: it_en_dev
+
+
+    # ST data
+    - name: de_fr_dev
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: de_fr_dev
+
+    - name: de_it_dev
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: de_it_dev
+
+    - name: fr_de_dev
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: fr_de_dev
+
+    - name: fr_it_dev
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: fr_it_dev
+
+    - name: it_de_dev
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: it_de_dev
+
+    - name: it_fr_dev
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: it_fr_dev
+
+
+  test:
+    # ASR data
+    - name: de_asr_test
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlASRDataset
+        split: de_en_test
+
+    - name: fr_asr_test
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlASRDataset
+        split: fr_en_test
+
+    - name: it_asr_test
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlASRDataset
+        split: it_en_test
+
+
+    # ST data
+    - name: de_fr_test
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: de_fr_test
+
+    - name: de_it_test
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: de_it_test
+
+    - name: fr_de_test
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: fr_de_test
+
+    - name: fr_it_test
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: fr_it_test
+
+    - name: it_de_test
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: it_de_test
+
+    - name: it_fr_test
+      dataset:
+        _target_: egs3.europarl.s2t1_new_conf.dataset.EuroparlSTDataset
+        split: it_fr_test
+
+  preprocessor:
+    _target_: egs3.europarl.s2t1_new_conf.dataset.OWSMTokenizeTransform
+    model_tag: espnet/owsm_v3.1_ebf
