+++ egs3/europarl/s2t1/path.sh
@@ -1,9 +1,8 @@
 #!/bin/bash
 
-# make sure we can use ffmpeg to load m4a audio
-export PATH=${PATH}:../../../tools/ffmpeg-release
+export HF_HOME=./hub
 export PYTHONPATH=${PYTHONPATH}:../../../
-. ../../../tools/activate_python.sh
+source ../../../tools/activate_python.sh
 
-# number of GPU, if we want to control from environment variable.
-export N_GPU=1
+# export WANDB_API_KEY=<>
+# wandb login
