+++ egs3/europarl/s2t1/train.py
@@ -3,26 +3,47 @@ import argparse
 import lightning as L
 import torch
 from hydra.utils import instantiate
-from omegaconf import OmegaConf
 from tqdm import tqdm
 
+from dask.distributed import WorkerPlugin, as_completed, get_worker
+
 from espnet3 import get_espnet_model, save_espnet_config
+from espnet3.utils.config import load_config_with_defaults
 from espnet3.preprocess import train_sentencepiece
 from espnet3.trainer import ESPnetEZLightningTrainer, LitESPnetModel
+from espnet3.parallel import get_client, set_parallel
 
 
-def load_line(path):
-    with open(path, "r") as f:
-        return [line.strip() for line in f.readlines()]
+class TokenizerPlugin(WorkerPlugin):
+    def __init__(self, dataset_config):
+        # remove valid/test config
+        dataset_config.pop("test")
+        self.dataset_config = dataset_config
+    
+    def setup(self, worker):
+        worker.dataset = instantiate(self.dataset_config)
 
 
 def train_tokenizer(config):
     config.dataset.preprocessor = None
     organizer = instantiate(config.dataset)
+    train_dataset_length = len(organizer.train)
+    del organizer
 
-    with open("train_text.txt", "w", encoding="utf-8") as f:
-        for example in tqdm(organizer.train):
-            f.write(example[1]["text"] + "\n")
+    def get_text(idx: int):
+        worker = get_worker()
+        dataset = worker.dataset
+        return dataset.train.get_text(idx)
+    
+    plugin = TokenizerPlugin(config.dataset)
+    set_parallel(config.parallel)
+
+    with get_client(plugin=plugin) as client, \
+        open("train_text.txt", "w", encoding="utf-8") as f:
+        futures = client.map(get_text, list(range(train_dataset_length)))
+        for future in tqdm(as_completed(futures), total=train_dataset_length):
+            text = future.result()
+            f.write(text + "\n")
             f.flush()
 
     train_sentencepiece(
@@ -36,7 +57,7 @@ def train_tokenizer(config):
 
 def main():
     parser = argparse.ArgumentParser()
-    parser.add_argument("--config", type=str, default="config.yaml")
+    parser.add_argument("--config", type=str, default="train.yaml")
     parser.add_argument(
         "--train_tokenizer", action="store_true", help="Train tokenizer before training"
     )
@@ -46,8 +67,7 @@ def main():
     args = parser.parse_args()
 
     # Load config
-    OmegaConf.register_new_resolver("load_line", load_line)
-    config = OmegaConf.load(args.config)
+    config = load_config_with_defaults(args.config)
 
     if args.train_tokenizer:
         print("==> Training tokenizer...")
@@ -58,11 +78,7 @@ def main():
         assert isinstance(config.seed, int), "seed should be an integer"
         L.seed_everything(config.seed)
 
-    # Set parallel config
-    # set_parallel(config.parallel)
-
-    # remove normalization configs for collect stats,
-    # as stats file collect_stats won't be available before processing.
+    # Prepare for collect_stats
     normalize = None
     normalize_conf = None
     if args.collect_stats:
