+++ egs3/europarl/s2t1/train.yaml
@@ -1,0 +1,1 @@
+conf/owsm_finetune.yaml
