+++ egs3/europarl/s2t1/train_gpu2.yaml
@@ -1,0 +1,1 @@
+conf/owsm_finetune_gpu2.yaml
