--- egs3/europarl/s2t1_new_conf/conf/owsm_finetune_gpu2.yaml
@@ -1,9 +1,0 @@
-# load default config
-defaults:
-  - dataset.yaml
-  - conf/default.yaml
-  - conf/owsm_finetune.yaml
-
-recipedir: .
-num_device: 2
-expdir: ${recipedir}/exp/train_ctc_gpu2
