--- egs3/europarl/s2t1_new_conf/evaluate.yaml
@@ -1,1 +1,0 @@
-conf/inference_beam1.yaml
