--- egs3/europarl/s2t1_new_conf/path.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-export HF_HOME=<>
-export PYTHONPATH=${PYTHONPATH}:../../../
-source ../../../tools/activate_python.sh
