--- egs3/europarl/s2t1_new_conf/train.yaml
@@ -1,1 +1,0 @@
-conf/train_ctc.yaml
